pyproject.toml
[project]
name = "fitness-analytics"
version = "0.1.0"
description = "Personal fitness intelligence platform — Strava + Hevy → DuckDB → SQLMesh → Evidence.dev"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "dlt[duckdb]>=1.0,<2.0",
    "sqlmesh[web]>=0.90,<1.0",
    "dagster>=1.6,<2.0",
    "dagster-webserver>=1.6,<2.0",
    "python-dotenv>=1.0,<2.0",
    "requests>=2.31,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "ipython>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
